@@ -18,14 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   shared_preferences: ^2.0.15
-  provider: ^6.0.4
-  flutter_bloc: ^8.1.1
+
 
 
   # The following adds the Cupertino Icons font to your application.
